requirements.txt
# Core ML/CV Dependencies
tensorflow==2.15.0
keras==2.15.0
deepface==0.0.75
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.2
opencv-python==4.8.1.78
ml-dtypes==0.2.0

# Streamlit & UI
streamlit==1.28.0
matplotlib==3.7.3
plotly==5.18.0

# LangChain Stack
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-groq==0.1.3
langsmith==0.1.51

# Utility Libraries
requests==2.31.0
python-dotenv==1.0.1
reportlab==4.0.4
youtube-search-python==1.6.6
Pillow==9.5.0
protobuf==4.25.3